refactor: change user card
