+++ src/components/UserCard.css
@@ -1,0 +1,94 @@
+.form {
+  width: 100%;
+  margin-top: 5vw;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  justify-content: center;
+  gap: 2vw;
+  padding-right: 40vw;
+}
+
+.form__label {
+  display: flex;
+  gap: 1vw;
+  align-items: center;
+  font-size: 1vw;
+}
+
+.form__btn {
+  width: 15vw;
+  height: 2vw;
+}
+
+input {
+  outline: none;
+  border: 0.1vw solid rgb(94, 93, 93);
+  font-size: 1vw;
+}
+
+#birthday,
+#name,
+#pet {
+  width: 15vw;
+  height: 2vw;
+}
+
+.form-card {
+  width: 25vw;
+  padding: 0.3vw;
+  display: flex;
+  gap: 1vw;
+  border: 1px solid rgb(77, 75, 75);
+}
+
+.form-card__about {
+  display: flex;
+  flex-direction: column;
+  gap: 0.7vw;
+}
+
+.form-card__img {
+  width: 10vw;
+  height: 10vw;
+}
+
+.form-card__name {
+  text-transform: uppercase;
+  letter-spacing: 0.1vw;
+  font-weight: 600;
+  font-size: 1.2vw;
+}
+
+.card-list {
+  margin: 5vw 0;
+}
+
+.card-list__inner {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2vw;
+}
+
+.message {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 300px;
+  color: white;
+  font-size: 20px;
+  text-transform: uppercase;
+  font-weight: 600;
+  text-align: center;
+  transform: translate(-50%, -50%);
+  background-color: rgb(36, 189, 9);
+}
+
+.error-message {
+  color: rgb(214, 7, 7);
+  font-weight: 600;
+  font-size: 0.9vw;
+  margin-top: -2vw;
+  margin-bottom: -0.9vw;
+  margin-right: 1.5vw;
+}
