+++ src/components/UserCard.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { FormProps } from '../types/interfaces';
+import './UserCard.css';
 
 type CardType = {
   card: FormProps;
