+++ src/pages/Forms.tsx
@@ -1,14 +1,35 @@
 import React, { Component } from 'react';
 import UserCardsList from '../components/UserCardList/UserCardsList';
-import Form from '../components/Form/Form';
+import { FormProps } from '../types/interfaces';
+import { FormAddCard } from '../components/FormAddCard/FormAddCard';
 
-export default class Forms extends Component {
+type Props = object;
+
+type State = {
+  data: Array<FormProps>;
+};
+
+export default class Forms extends Component<Props, State> {
+  constructor(props: object) {
+    super(props);
+    this.state = { data: [] };
+    this.handleData = this.handleData.bind(this);
+  }
+
+  handleData(dataCard: FormProps) {
+    this.setState((prevState) => ({
+      data: [...prevState.data, dataCard],
+    }));
+  }
+
   render() {
+    const { data } = this.state;
+
     return (
       <div className="wrapper">
         <h1 style={{ color: '#b1e926' }}>Form</h1>
-        <Form />
-        <UserCardsList />
+        <FormAddCard onSubmit={this.handleData} />
+        <UserCardsList cards={data} />
       </div>
     );
   }
